support scikit-learn==1.8.0 (#1219)
* fix style after updating black


* udate CI to test test scikit-learn==1.8.0


* fix ir_version correspondance with opset


* fix unit test for 1.8.0


* fix ir_version


* use opset 22


* fix more unit tests


* fix more unittest


* fix more unittest


* final changes


* remove unnecessary parameters'


* fix more examples


* fix examples


* remove python 3.9


* ci


* ci


* fix ci


* fix ndonnx


* fix example


---------

+++ skl2onnx/operator_converters/linear_classifier.py
@@ -53,11 +53,11 @@ def convert_sklearn_linear_classifier(
         # fit_intercept = False
         intercepts = [0.0] * number_of_classes if number_of_classes != 2 else [0.0]
     else:
-        intercepts = op.intercept_.tolist()
+        intercepts = op.intercept_.astype(float).tolist()
 
     if number_of_classes == 2:
-        coefficients = list(map(lambda x: -1 * x, coefficients)) + coefficients
-        intercepts = list(map(lambda x: -1 * x, intercepts)) + intercepts
+        coefficients = list(map(lambda x: -1.0 * x, coefficients)) + coefficients
+        intercepts = list(map(lambda x: -1.0 * x, intercepts)) + intercepts
 
     multi_class = 0
     use_ovr = False
